+++ packages/css-color-parser/src/color-data.ts
@@ -513,11 +513,33 @@ function reducePrecision(x: number, precision = 7): number {
 }
 
 export function colorDataFitsRGB_Gamut(x: ColorData): boolean {
-	const copy = JSON.parse(JSON.stringify(x)) as ColorData;
+	const copy: ColorData = {
+		...x,
+		channels: [
+			...x.channels,
+		],
+	};
 
 	copy.channels = convertPowerlessComponentsToZeroValuesForDisplay(copy.channels, copy.colorNotation);
 	const srgb = colorDataTo(copy, ColorNotation.RGB);
 	if (!srgb.channels.find((y) => y < -0.00001 || y > 1.00001)) {
+		return true;
+	}
+
+	return false;
+}
+
+export function colorDataFitsDisplayP3_Gamut(x: ColorData): boolean {
+	const copy: ColorData = {
+		...x,
+		channels: [
+			...x.channels,
+		],
+	};
+
+	copy.channels = convertPowerlessComponentsToZeroValuesForDisplay(copy.channels, copy.colorNotation);
+	const displayP3 = colorDataTo(copy, ColorNotation.Display_P3);
+	if (!displayP3.channels.find((y) => y < -0.00001 || y > 1.00001)) {
 		return true;
 	}