+++ plugin-packs/postcss-preset-env/src/lib/ids-by-execution-order.mjs
@@ -60,4 +60,7 @@ export default [
 	'custom-properties',
 
 	'cascade-layers',
+
+	'progressive-custom-properties',
+	'gamut-mapping',
 ];