+++ plugin-packs/postcss-preset-env/src/test/lib/list-features/no-options.mjs
@@ -28,6 +28,12 @@ assert.deepStrictEqual(
 			},
 			id: 'any-link-pseudo-class',
 		},
+		{
+			browsers: ['> 0%'],
+			vendors_implementations: undefined,
+			pluginOptions: {},
+			id: 'progressive-custom-properties',
+		},
 	],
 );