+++ plugins/postcss-progressive-custom-properties/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changes to PostCSS Progressive Custom Properties
 
+### Unreleased (patch)
+
+- Add support for css nesting
+- Improve grouping of `@supports` rules
+- Reduce the size of the generated `@supports` rules for values that contain `var()`
+
 ### 3.0.1
 
 _September 18, 2023_